clean up code (subscriber)
