+++ app/src/main/java/org/positive/daymotion/presentation/base/BaseViewModel.kt
@@ -5,18 +5,19 @@ import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import io.reactivex.rxjava3.disposables.CompositeDisposable
 import org.positive.daymotion.common.SingleLiveEvent
-import org.positive.daymotion.presentation.base.subscope.*
+import org.positive.daymotion.presentation.base.subscriber.*
 
 abstract class BaseViewModel constructor(
-    private val disposables: CompositeDisposable = CompositeDisposable(),
-    private val _isLoading: MutableLiveData<Boolean> = MutableLiveData<Boolean>(),
-    private val _showErrorMessageEvent: SingleLiveEvent<String> = SingleLiveEvent()
+    disposables: CompositeDisposable = CompositeDisposable(),
+    isLoading: MutableLiveData<Boolean> = MutableLiveData<Boolean>(),
 ) : ViewModel(),
-    SingleSubscriber by SingleSubscriberImpl(disposables, _isLoading),
-    MaybeSubscriber by MaybeSubscriberImpl(disposables, _isLoading),
-    CompletableSubscriber by CompletableSubscriberImpl(disposables, _isLoading) {
+    SingleSubscriber by SingleSubscriberImpl(disposables, isLoading),
+    MaybeSubscriber by MaybeSubscriberImpl(disposables, isLoading),
+    CompletableSubscriber by CompletableSubscriberImpl(disposables, isLoading) {
 
-    val isLoading: LiveData<Boolean> = _isLoading
+    val isLoading: LiveData<Boolean> = isLoading
+
+    private val _showErrorMessageEvent: SingleLiveEvent<String> = SingleLiveEvent()
     val showErrorMessageEvent: LiveData<String> = _showErrorMessageEvent
 
     protected fun showErrorMessage(message: String) {
