+++ app/src/main/java/org/positive/daymotion/presentation/base/subscriber/CompletableSubscribeScope.kt
@@ -1,4 +1,4 @@
-package org.positive.daymotion.presentation.base.subscope
+package org.positive.daymotion.presentation.base.subscriber
 
 import io.reactivex.rxjava3.core.Completable
 import io.reactivex.rxjava3.disposables.CompositeDisposable
