+++ app/src/main/java/org/positive/daymotion/presentation/base/subscriber/MaybeSubscriber.kt
@@ -1,37 +1,37 @@
-package org.positive.daymotion.presentation.base.subscope
+package org.positive.daymotion.presentation.base.subscriber
 
 import androidx.lifecycle.MutableLiveData
 import io.reactivex.rxjava3.android.schedulers.AndroidSchedulers
-import io.reactivex.rxjava3.core.Completable
+import io.reactivex.rxjava3.core.Maybe
 import io.reactivex.rxjava3.disposables.CompositeDisposable
 import io.reactivex.rxjava3.schedulers.Schedulers
 
-interface CompletableSubscriber {
+interface MaybeSubscriber {
 
-    fun Completable.autoDispose(block: CompletableSubscribeScope.() -> Unit)
+    fun <T> Maybe<T>.autoDispose(block: MaybeSubscribeScope<T>.() -> Unit)
 
-    fun Completable.backgroundCompose(): Completable
+    fun <T> Maybe<T>.backgroundCompose(): Maybe<T>
 
-    fun Completable.apiLoadingCompose(): Completable
+    fun <T> Maybe<T>.apiLoadingCompose(): Maybe<T>
 }
 
-class CompletableSubscriberImpl(
+class MaybeSubscriberImpl(
     private val disposables: CompositeDisposable,
     private val _isLoading: MutableLiveData<Boolean>
-) : CompletableSubscriber {
+) : MaybeSubscriber {
 
-    override fun Completable.autoDispose(block: CompletableSubscribeScope.() -> Unit) {
-        val scope = CompletableSubscribeScope(this, disposables)
+    override fun <T> Maybe<T>.autoDispose(block: MaybeSubscribeScope<T>.() -> Unit) {
+        val scope = MaybeSubscribeScope(this, disposables)
         scope.block()
         scope.subscribe()
     }
 
-    override fun Completable.backgroundCompose(): Completable = compose {
+    override fun <T> Maybe<T>.backgroundCompose(): Maybe<T> = compose {
         it.subscribeOn(Schedulers.io())
             .observeOn(AndroidSchedulers.mainThread())
     }
 
-    override fun Completable.apiLoadingCompose(): Completable = compose {
+    override fun <T> Maybe<T>.apiLoadingCompose(): Maybe<T> = compose {
         it.subscribeOn(Schedulers.io())
             .observeOn(AndroidSchedulers.mainThread())
             .doOnSubscribe { _isLoading.value = true }
