+++ app/src/main/java/org/positive/daymotion/presentation/base/subscriber/SingleSubscribeScope.kt
@@ -1,15 +1,14 @@
-package org.positive.daymotion.presentation.base.subscope
+package org.positive.daymotion.presentation.base.subscriber
 
-import io.reactivex.rxjava3.core.Maybe
+import io.reactivex.rxjava3.core.Single
 import io.reactivex.rxjava3.disposables.CompositeDisposable
 
-class MaybeSubscribeScope<T>(
-    private val source: Maybe<T>,
+class SingleSubscribeScope<T>(
+    private val source: Single<T>,
     private val disposables: CompositeDisposable
 ) {
     private var onSuccess: (T) -> Unit = {}
     private var onError: (Throwable) -> Unit = {}
-    private var onComplete: () -> Unit = {}
 
     fun success(onSuccess: (T) -> Unit) {
         this.onSuccess = onSuccess
@@ -19,11 +18,7 @@ class MaybeSubscribeScope<T>(
         this.onError = onError
     }
 
-    fun complete(onComplete: () -> Unit) {
-        this.onComplete = onComplete
-    }
-
     fun subscribe() {
-        disposables.add(source.subscribe(onSuccess, onError, onComplete))
+        disposables.add(source.subscribe(onSuccess, onError))
     }
 }
